+++ src/Concerns/PerformsCharges.php
@@ -23,6 +23,7 @@ trait PerformsCharges
     public function charge($amount, $paymentMethod, array $options = [])
     {
         $options = array_merge([
+            'automatic_tax' => $this->automaticTaxPayload(),
             'confirmation_method' => 'automatic',
             'confirm' => true,
             'currency' => $this->preferredCurrency(),
@@ -71,6 +72,7 @@ trait PerformsCharges
     {
         $payload = array_filter([
             'allow_promotion_codes' => $this->allowPromotionCodes,
+            'automatic_tax' => $this->automaticTaxPayload(),
             'discounts' => $this->checkoutDiscounts(),
             'line_items' => Collection::make((array) $items)->map(function ($item, $key) {
                 if (is_string($key)) {
@@ -83,6 +85,7 @@ trait PerformsCharges
 
                 return $item;
             })->values()->all(),
+            'tax_id_collection' => $this->collectsTaxIds,
         ]);
 
         return Checkout::create($this, array_merge($payload, $sessionOptions), $customerOptions);