+++ src/Subscription.php
@@ -9,6 +9,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 use InvalidArgumentException;
+use Laravel\Cashier\Concerns\HandlesTaxes;
 use Laravel\Cashier\Concerns\InteractsWithPaymentBehavior;
 use Laravel\Cashier\Concerns\Prorates;
 use Laravel\Cashier\Database\Factories\SubscriptionFactory;
@@ -22,6 +23,7 @@ use Stripe\Subscription as StripeSubscription;
  */
 class Subscription extends Model
 {
+    use HandlesTaxes;
     use HasFactory;
     use InteractsWithPaymentBehavior;
     use Prorates;
@@ -789,6 +791,7 @@ class Subscription extends Model
     protected function getSwapOptions(Collection $items, array $options = [])
     {
         $payload = [
+            'automatic_tax' => $this->automaticTaxPayload(),
             'items' => $items->values()->all(),
             'payment_behavior' => $this->paymentBehavior(),
             'proration_behavior' => $this->prorateBehavior(),
@@ -830,14 +833,15 @@ class Subscription extends Model
             throw SubscriptionUpdateFailure::duplicatePrice($this, $price);
         }
 
-        $stripeSubscriptionItem = $this->owner->stripe()->subscriptionItems->create(array_merge([
-            'subscription' => $this->stripe_id,
-            'price' => $price,
-            'quantity' => $quantity,
-            'tax_rates' => $this->getPriceTaxRatesForPayload($price),
-            'payment_behavior' => $this->paymentBehavior(),
-            'proration_behavior' => $this->prorateBehavior(),
-        ], $options));
+        $stripeSubscriptionItem = $this->owner->stripe()->subscriptionItems
+            ->create(array_filter(array_merge([
+                'subscription' => $this->stripe_id,
+                'price' => $price,
+                'quantity' => $quantity,
+                'tax_rates' => $this->getPriceTaxRatesForPayload($price),
+                'payment_behavior' => $this->paymentBehavior(),
+                'proration_behavior' => $this->prorateBehavior(),
+            ], $options)));
 
         $this->items()->create([
             'stripe_id' => $stripeSubscriptionItem->id,
@@ -1128,6 +1132,7 @@ class Subscription extends Model
     public function upcomingInvoice(array $options = [])
     {
         return $this->owner->upcomingInvoice(array_merge([
+            'automatic_tax' => $this->automaticTaxPayload(),
             'subscription' => $this->stripe_id,
         ], $options));
     }